Split the CmShell into several files and refac
Make changes regarding new linter rules and change how the cmshell stops.
